refactor : Winners 메소드 Cars로 이동
