+++ src/test/java/racingcar/model/CarsTest.java
@@ -1,12 +1,18 @@
 package racingcar.model;
 
 import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
+
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.List;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 class CarsTest {
 
+	int expectMaxPosition = 4;
 	Cars cars;
 
 	@Test
